For #76, rethink typed optional args of options.
With this change, when we encounter an option that accepts an optional
argument, then if the next argument does not "look like" an option, and
the argument can be converted to the desired type, then we'll treat the
arg as the argument of the option. If the next argument "looks like" an
option, but the option's argument is a number type, then we'll treat the
argument as the numeric value of the option. (This allows for things like
--factor -2.)

If neither of the above scenarios applies, then the option is considered
not to have an argument.