Make AbstractOptionSpec for kotlin.
Kotlin does not allow to use classes in package scope
even when they have public methods.
When a class in package scope has public methods
these bleed into the public space anyway.
So make the class public but let the constructors
be package protected so it may only be instantiated
from within the package.