Make long opt ending with equals have empty value
With a "delete" option that requires an argument and an "add" option
that does not, JOpt Simple parses the arguments ["--delete=", "--add"]
as the option "delete" with the value "--add", and no "add" option
detected.  This is incorrect.  It should parse it as the option "delete"
with the value "" (i.e., an empty string), and the option "add"
detected.  Make JOpt Simple treat an option that ends with an equals
sign as an option with an empty value.

As of May 5, 2017, on macOS Sierra 10.12.4, the getopt-long example at

  https://www.gnu.org/software/libc/manual/html_node/Getopt-Long-Option-Example.html

behaves in the expected way ("d" and "a" are the short forms of "delete"
and "add", respectively):

  % ./main --delete= --add
  option -d with value `'
  option -a